requirements.txt
anyio==4.0.0
blessed==1.20.0
certifi==2023.7.22
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
exceptiongroup==1.1.3
fonttools==4.45.1
h11==0.14.0
h2o-wave==1.0.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
inquirer==3.1.3
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-editor==1.0.4
readchar==4.0.5
six==1.16.0
sniffio==1.3.0
starlette==0.31.1
typing_extensions==4.8.0
uvicorn==0.23.2
wcwidth==0.2.8
